eliminate "type" parameter
"type" is only a parameter for the Poderopedia API (for which it is retained). Otherwise, "type" was implemented as a meta-parameter, to handle the case where an API was known to only respond with people or organizations. (It was originally implemented thinking that no API merged people and organization into one endpoint; this assumption did not hold.) Since it is the only meta-parameter, it's being eliminated.